requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.3.0
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
asttokens=3.0.0=pyhd8ed1ab_1
blessed=1.21.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.4.26=hbd8a1cb_0
certifi=2025.4.26=pypi_0
charset-normalizer=3.4.2=pypi_0
clip=0.2.0=pypi_0
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.2=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.14=py310hf71b8c6_0
decorator=5.2.1=pyhd8ed1ab_0
exceptiongroup=1.3.0=pyhd8ed1ab_0
executing=2.2.0=pyhd8ed1ab_0
fonttools=4.58.1=pypi_0
gpustat=1.1.1=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.7.0=pyhe01879c_1
ipykernel=6.29.5=pyh3099207_0
ipython=8.37.0=pyh8f84b5b_0
jedi=0.19.2=pyhd8ed1ab_1
joblib=1.5.1=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.8.1=pyh31011fe_0
kiwisolver=1.4.8=pypi_0
krb5=1.21.3=h143b758_0
ld_impl_linux-64=2.40=h12ee557_0
libedit=3.1.20230828=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc=15.1.0=h767d61c_2
libgcc-ng=15.1.0=h69a702a_2
libgomp=15.1.0=h767d61c_2
libsodium=1.0.20=h4ab18f5_0
libstdcxx=15.1.0=h8f9b012_2
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
libxcb=1.17.0=h9b100fa_0
llvmlite=0.43.0=pypi_0
matplotlib=3.6.2=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
numba=0.60.0=pypi_0
numpy=1.23.4=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-ml-py=12.575.51=pypi_0
openssl=3.5.0=h7b32b05_1
packaging=25.0=pyh29332c3_1
pandas=1.5.2=pypi_0
parso=0.8.4=pyhd8ed1ab_1
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=9.5.0=pypi_0
pip=25.1=pyhc872135_2
platformdirs=4.3.8=pyhe01879c_0
pot=0.8.2=pypi_0
prompt-toolkit=3.0.51=pyha770c72_0
psutil=7.0.0=py310ha75aee5_0
pthread-stubs=0.3=h0ce48e5_1
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pygments=2.19.1=pyhd8ed1ab_0
pynndescent=0.5.13=pypi_0
pyparsing=3.2.3=pypi_0
python=3.10.16=he870216_1
python-dateutil=2.9.0.post0=pyhff2d567_1
python_abi=3.10=2_cp310
pytz=2025.2=pypi_0
pyzmq=26.4.0=py310h71f11fc_0
readline=8.2=h5eee18b_0
requests=2.32.3=pypi_0
scikit-learn=1.1.3=pypi_0
scipy=1.13.1=pypi_0
setuptools=78.1.1=py310h06a4308_0
six=1.17.0=pyhd8ed1ab_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.3=pyhd8ed1ab_1
threadpoolctl=3.6.0=pypi_0
tk=8.6.14=h993c535_1
torch=1.13.1=pypi_0
torchvision=0.14.1=pypi_0
tornado=6.5.1=py310ha75aee5_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
typing_extensions=4.13.2=pyh29332c3_0
tzdata=2025b=h04d1e81_0
umap=0.1.1=pypi_0
umap-learn=0.5.3=pypi_0
urllib3=2.4.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
wheel=0.45.1=py310h06a4308_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
zeromq=4.3.5=h3b0a872_7
zipp=3.22.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_1
